Revert "[core/gpr] move subprocess to gpr (#33870)"
This reverts commit f10a8e3418e51230b0af86e880f82c8325a68abe.
