+++ grpc.gyp
@@ -1504,6 +1504,8 @@
         'test/core/util/passthru_endpoint.cc',
         'test/core/util/resolve_localhost_ip46.cc',
         'test/core/util/slice_splitter.cc',
+        'test/core/util/subprocess_posix.cc',
+        'test/core/util/subprocess_windows.cc',
         'test/core/util/tracer_util.cc',
         'test/cpp/microbenchmarks/helpers.cc',
       ],
@@ -1644,8 +1646,6 @@
         'grpc_test_util',
       ],
       'sources': [
-        'src/core/lib/gpr/subprocess_posix.cc',
-        'src/core/lib/gpr/subprocess_windows.cc',
         'test/core/end2end/data/client_certs.cc',
         'test/core/end2end/data/server1_cert.cc',
         'test/core/end2end/data/server1_key.cc',
@@ -1659,6 +1659,8 @@
         'test/core/util/passthru_endpoint.cc',
         'test/core/util/resolve_localhost_ip46.cc',
         'test/core/util/slice_splitter.cc',
+        'test/core/util/subprocess_posix.cc',
+        'test/core/util/subprocess_windows.cc',
         'test/core/util/tracer_util.cc',
         'test/cpp/util/byte_buffer_proto_helper.cc',
         'test/cpp/util/create_test_channel.cc',
