+++ test/core/http/httpcli_test_util.cc
@@ -32,8 +32,8 @@
 #include <grpc/support/time.h>
 
 #include "src/core/lib/config/config_vars.h"
-#include "src/core/lib/gpr/subprocess.h"
 #include "test/core/util/port.h"
+#include "test/core/util/subprocess.h"
 
 namespace grpc_core {
 namespace testing {
