+++ test/core/util/subprocess_windows.cc
@@ -31,9 +31,9 @@
 #include <grpc/support/log.h>
 
 #include "src/core/lib/gpr/string.h"
-#include "src/core/lib/gpr/subprocess.h"
 #include "src/core/lib/gprpp/crash.h"
 #include "src/core/lib/gprpp/tchar.h"
+#include "test/core/util/subprocess.h"
 
 struct gpr_subprocess {
   PROCESS_INFORMATION pi;
