+++ test/cpp/util/BUILD
@@ -80,7 +80,6 @@ grpc_cc_library(
     ],
     deps = [
         "//:grpc++",
-        "//:subprocess",
         "//test/core/end2end:ssl_test_data",
         "//test/core/util:grpc_test_util",
         "//test/core/util:grpc_test_util_base",
@@ -115,7 +114,6 @@ grpc_cc_library(
     ],
     deps = [
         "//:grpc++_unsecure",
-        "//:subprocess",
         "//test/core/util:grpc_test_util_base",
         "//test/core/util:grpc_test_util_unsecure",
     ],
